+++ deploy.sh
@@ -12,6 +12,13 @@ if [[ $rc -ne 0 ]]; then
     echo "Build failed with exit status $rc! Aborting deploy."
     exit 1
 fi
+if [ -d $BUILD_DIR/queries ]; then
+    echo "No queries directory found in $BUILD_DIR."
+    echo "Creating queries directory in $BUILD_DIR..."
+    mkdir $BUILD_DIR/queries
+fi
+echo "Copying SQL queries to $BUILD_DIR/queries/sql..."
+cp -r $SRC_DIR/queries/sql $BUILD_DIR/queries/sql
 echo "Deploying via serverless framework..."
 cd $BASE_DIR
 serverless deploy 2> /dev/null